refactor: add --discreetly flag make display discreet optional
