+++ cli/src/commands/account.rs
@@ -22,9 +22,11 @@ use anyhow::{anyhow, bail, Result};
 use clap::Parser;
 use colored::Colorize;
 use core::str::FromStr;
+use crossterm::ExecutableCommand;
 use rand::SeedableRng;
 use rand_chacha::ChaChaRng;
 use rayon::prelude::*;
+use std::io::{Read, Write};
 
 type Network = snarkvm::prelude::Testnet3;
 
@@ -39,13 +41,16 @@ pub enum Account {
         /// Try until an address with the vanity string is found
         #[clap(short = 'v', long)]
         vanity: Option<String>,
+        /// Print sensitive information(such as private key) discreetly to an alternate screen
+        #[clap(long)]
+        discreet: bool,
     },
 }
 
 impl Account {
     pub fn parse(self) -> Result<String> {
         match self {
-            Self::New { seed, vanity } => {
+            Self::New { seed, vanity, discreet } => {
                 // Ensure only the seed or the vanity string is specified.
                 if seed.is_some() && vanity.is_some() {
                     bail!("Cannot specify both the '--seed' and '--vanity' flags");
@@ -53,18 +58,18 @@ impl Account {
 
                 // Generate a vanity account.
                 if let Some(vanity) = vanity {
-                    Self::new_vanity(&vanity)
+                    Self::new_vanity(&vanity, discreet)
                 }
                 // Default to generating a normal account, with an optional seed.
                 else {
-                    Self::new_seeded(seed)
+                    Self::new_seeded(seed, discreet)
                 }
             }
         }
     }
 
     /// Generates a new Aleo account with the given vanity string.
-    fn new_vanity(vanity: &str) -> Result<String> {
+    fn new_vanity(vanity: &str, discreet: bool) -> Result<String> {
         // A closure to generate a new Aleo account.
         let sample_account = || snarkos_account::Account::<Network>::new(&mut rand::thread_rng());
 
@@ -112,7 +117,22 @@ impl Account {
             // Return the result if a candidate was found.
             if let Some(account) = account {
                 println!(); // Add a newline for formatting.
-                return Ok(account.to_string());
+                if !discreet {
+                    return Ok(account.to_string());
+                }
+                display_string_discreetly(
+                    &format!("{:>12}  {}", "Private Key".cyan().bold(), account.private_key()),
+                    "### Do not share or lose this private key! Press any key to complete. ###",
+                )
+                .unwrap();
+                let account_info = format!(
+                    " {:>12}  {}\n {:>12}  {}",
+                    "View Key".cyan().bold(),
+                    account.view_key(),
+                    "Address".cyan().bold(),
+                    account.address()
+                );
+                return Ok(account_info);
             } else {
                 let rate = ITERATIONS / timer.elapsed().as_millis();
                 let rate = format!("[{rate} a/ms]");
@@ -122,7 +142,7 @@ impl Account {
     }
 
     /// Generates a new Aleo account with an optional seed.
-    fn new_seeded(seed: Option<String>) -> Result<String> {
+    fn new_seeded(seed: Option<String>, discreet: bool) -> Result<String> {
         // Recover the seed.
         let seed = match seed {
             // Recover the field element deterministically.
@@ -138,8 +158,44 @@ impl Account {
         // Construct the account.
         let account = snarkos_account::Account::<Network>::try_from(private_key)?;
         // Print the new Aleo account.
-        Ok(account.to_string())
+        if !discreet {
+            return Ok(account.to_string());
+        }
+        display_string_discreetly(
+            &format!("{:>12}  {}", "Private Key".cyan().bold(), account.private_key()),
+            "### Do not share or lose this private key! Press any key to complete. ###",
+        )
+        .unwrap();
+        let account_info = format!(
+            " {:>12}  {}\n {:>12}  {}",
+            "View Key".cyan().bold(),
+            account.view_key(),
+            "Address".cyan().bold(),
+            account.address()
+        );
+        Ok(account_info)
     }
+}
+
+// Print the string to an alternate screen, so that the string won't been printed to the terminal.
+fn display_string_discreetly(discreet_string: &str, continue_message: &str) -> Result<()> {
+    use crossterm::{
+        style::Print,
+        terminal::{EnterAlternateScreen, LeaveAlternateScreen},
+    };
+    let mut stdout = std::io::stdout();
+    stdout.execute(EnterAlternateScreen)?;
+    // print msg on the alternate screen
+    stdout.execute(Print(format!("{discreet_string}\n{continue_message}")))?;
+    stdout.flush()?;
+    wait_for_keypress();
+    stdout.execute(LeaveAlternateScreen)?;
+    Ok(())
+}
+
+fn wait_for_keypress() {
+    let mut single_key = [0u8];
+    std::io::stdin().read_exact(&mut single_key).unwrap();
 }
 
 #[cfg(test)]
@@ -151,7 +207,7 @@ mod tests {
     #[test]
     fn test_new() {
         for _ in 0..3 {
-            let account = Account::New { seed: None, vanity: None };
+            let account = Account::New { seed: None, vanity: None, discreet: false };
             assert!(account.parse().is_ok());
         }
     }
@@ -177,7 +233,7 @@ mod tests {
         );
 
         let vanity = None;
-        let account = Account::New { seed, vanity };
+        let account = Account::New { seed, vanity, discreet: false };
         let actual = account.parse().unwrap();
         assert_eq!(expected, actual);
     }
@@ -203,7 +259,7 @@ mod tests {
         );
 
         let vanity = None;
-        let account = Account::New { seed, vanity };
+        let account = Account::New { seed, vanity, discreet: false };
         let actual = account.parse().unwrap();
         assert_eq!(expected, actual);
     }
